move find day offset to utils
